Successful launch 4.
Modified Drive module with Darrel's bang-bang changes.
Switched bang-bang thresholds, added deadband.
Changed navigation to use percent speed with motors.

+++ include/Drive.h
@@ -83,7 +83,7 @@ void Drive_backward(uint8_t speed);
  * @author David Goodman
  * @date 2013.03.30 
  */
-void Drive_forwardHeading(float speed, uint16_t angle);
+void Drive_forwardHeading(uint8_t speed, uint16_t angle);
 
 /**
  * Function: Drive_pivot
